wtech_ai_cnn/requirements.txt
tensorflow==2.14.0
pandas==2.1.3
scikit-learn==1.3.2
Pillow==10.1.0
psutil==5.9.5
ipykernel

fastapi
uvicorn
pydantic